+++ src/senaite/core/upgrade/v02_06_000.py
@@ -89,6 +89,8 @@ REMOVE_AT_TYPES = [
     "BatchLabels",
     "AnalysisCategory",
     "AnalysisCategories",
+    "AttachmentType",
+    "AttachmentTypes",
 ]
 
 CONTENT_ACTIONS = [
@@ -1533,10 +1535,6 @@ def move_instrumentlocations(tool):
     """Move instrument locations to senaite setup folder
     """
 
-    # run required import steps
-    tool.runImportStepFromProfile(profile, "typeinfo")
-    tool.runImportStepFromProfile(profile, "workflow")
-
     # get the old container
     origin = api.get_setup().get("instrumentlocations")
     if not origin:
@@ -1569,10 +1567,6 @@ def move_samplecontainers(tool):
     """Move sample containers to senaite setup folder
     """
 
-    # run required import steps
-    tool.runImportStepFromProfile(profile, "typeinfo")
-    tool.runImportStepFromProfile(profile, "workflow")
-
     # get the old container
     origin = api.get_setup().get("sample_containers")
     if not origin:
@@ -1598,6 +1592,53 @@ def move_samplecontainers(tool):
         logger.warn("Cannot remove {}. Is not empty".format(origin))
 
     logger.info("Move Sample Containers [DONE]")
+
+
+@upgradestep(product, version)
+def migrate_attachmenttypes_to_dx(tool):
+    """Converts existing attachment types to Dexterity
+    """
+    logger.info("Convert AttachmentTypes to Dexterity ...")
+
+    # ensure old AT types are flushed first
+    remove_at_portal_types(tool)
+
+    # run required import steps
+    tool.runImportStepFromProfile(profile, "typeinfo")
+    tool.runImportStepFromProfile(profile, "workflow")
+
+    # get the old container
+    origin = api.get_setup().get("bika_attachmenttypes")
+    if not origin:
+        # old container is already gone
+        return
+
+    # get the destination container
+    destination = get_setup_folder("attachmenttypes")
+
+    # un-catalog the old container
+    uncatalog_object(origin)
+
+    # Mapping from schema field name to a tuple of
+    # (accessor, target field name, default value)
+    schema_mapping = {
+        "title": ("Title", "title", ""),
+        "description": ("Description", "description", ""),
+    }
+
+    # migrate the contents from the old AT container to the new one
+    migrate_to_dx("AttachmentType", origin, destination, schema_mapping)
+
+    # copy snapshots for the container
+    copy_snapshots(origin, destination)
+
+    # remove old AT folder
+    if len(origin) == 0:
+        delete_object(origin)
+    else:
+        logger.warn("Cannot remove {}. Is not empty".format(origin))
+
+    logger.info("Convert AttachmentTypes to Dexterity [DONE]")
 
 
 def update_content_actions(tool):